refactor: reorganize event components to use common structure
- Updated event-related components to utilize a common structure for better maintainability.
- Moved components from specific directories to a shared 'common' directory for consistency.
- Removed deprecated components and adjusted imports across various event pages for AU, CA, GB, and US.
- Introduced new dialog and layout components to enhance user experience and streamline event presentation.
