+++ src/app/[countryCode]/events/[state]/[eventSlug]/page.tsx
@@ -2,7 +2,7 @@ import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import sanitizeHtml from 'sanitize-html'
 
-import { EventDialogContent } from '@/components/app/pageEvents/components/eventDialogContent'
+import { EventDialogContent } from '@/components/app/pageEvents/common/eventDialogContent'
 import { UsEventsPageDialogDeeplinkLayout } from '@/components/app/pageEvents/us/eventsPageDialogDeeplinkLayout'
 import { PageProps } from '@/types'
 import { getEvent, getEvents } from '@/utils/server/builder/models/data/events'
