+++ src/app/ca/events/[state]/[eventSlug]/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from 'next/navigation'
 import sanitizeHtml from 'sanitize-html'
 
 import { CaEventsPageDialogDeeplinkLayout } from '@/components/app/pageEvents/ca/eventsPageDialogDeeplinkLayout'
-import { EventDialogContent } from '@/components/app/pageEvents/components/eventDialogContent'
+import { EventDialogContent } from '@/components/app/pageEvents/common/eventDialogContent'
 import { PageProps } from '@/types'
 import { getEvent, getEvents } from '@/utils/server/builder/models/data/events'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
