+++ src/app/ca/events/[state]/page.tsx
@@ -2,7 +2,7 @@ import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 
 import { CaEventsPageDialogDeeplinkLayout } from '@/components/app/pageEvents/ca/eventsPageDialogDeeplinkLayout'
-import { StateEventsDialogContent } from '@/components/app/pageEvents/components/stateEventsDialogContent'
+import { StateEventsDialogContent } from '@/components/app/pageEvents/common/stateEventsDialogContent'
 import { PageProps } from '@/types'
 import { getEvents } from '@/utils/server/builder/models/data/events'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
