+++ src/components/app/pageEvents/au/eventDialogContent.tsx
@@ -1,0 +1,3 @@
+export function AuEventDialogContent() {
+  return <div>AuEventDialogContent</div>
+}
