+++ src/components/app/pageEvents/common/eventDialogContent.tsx
@@ -8,11 +8,11 @@ import sanitizeHtml from 'sanitize-html'
 import { toast } from 'sonner'
 
 import { LoginDialogWrapper } from '@/components/app/authentication/loginDialogWrapper'
-import { EventDialogPhoneNumber } from '@/components/app/pageEvents/components/eventDialogPhoneNumber'
-import { EventDialogSocialLinks } from '@/components/app/pageEvents/components/eventDialogSocialLinks'
-import { GoogleMapsEmbedIFrame } from '@/components/app/pageEvents/components/eventGoogleMapsEmbedIframe'
-import { EventPhotosSlideshow } from '@/components/app/pageEvents/components/eventPhotosSlideshow'
-import { SuccessfulEventNotificationsSignup } from '@/components/app/pageEvents/components/successfulEventSignupDialog'
+import { EventDialogPhoneNumber } from '@/components/app/pageEvents/common/eventDialogPhoneNumber'
+import { EventDialogSocialLinks } from '@/components/app/pageEvents/common/eventDialogSocialLinks'
+import { GoogleMapsEmbedIFrame } from '@/components/app/pageEvents/common/eventGoogleMapsEmbedIframe'
+import { EventPhotosSlideshow } from '@/components/app/pageEvents/common/eventPhotosSlideshow'
+import { SuccessfulEventNotificationsSignup } from '@/components/app/pageEvents/common/successfulEventSignupDialog'
 import { handleCreateRsvpAction as _handleCreateRsvpAction } from '@/components/app/pageEvents/utils/createRsvpAction'
 import { Button } from '@/components/ui/button'
 import { NextImage } from '@/components/ui/image'
