+++ src/components/app/pageEvents/common/eventDialogContentLazyload.tsx
@@ -1,7 +1,7 @@
 import { lazy } from 'react'
 
 export const LazyEventDialogContent = lazy(() =>
-  import('@/components/app/pageEvents/components/eventDialogContent').then(m => ({
+  import('@/components/app/pageEvents/common/eventDialogContent').then(m => ({
     default: m.EventDialogContent,
   })),
 )
