+++ src/components/app/pageEvents/common/eventDialogContentSkeleton.tsx
