+++ src/components/app/pageEvents/common/eventPhotosSlideshow.tsx
