+++ src/components/app/pageEvents/common/eventsMap.tsx
@@ -8,7 +8,7 @@ import { MAP_PROJECTION_CONFIG } from '@/components/app/pageAdvocatesHeatmap/con
 import {
   StateEventsDialog,
   StateEventsDialogProps,
-} from '@/components/app/pageEvents/components/stateEventsDialog'
+} from '@/components/app/pageEvents/common/stateEventsDialog'
 import { EVENT_MAP_STATE_COORDS } from '@/components/app/pageEvents/utils/mapCoordinates'
 import { pluralize } from '@/utils/shared/pluralize'
 import {
