+++ src/components/app/pageEvents/common/stateEventsDialogContentLazyload.tsx
@@ -1,7 +1,7 @@
 import { lazy } from 'react'
 
 export const LazyStateEventsDialogContent = lazy(() =>
-  import('@/components/app/pageEvents/components/stateEventsDialogContent').then(m => ({
+  import('@/components/app/pageEvents/common/stateEventsDialogContent').then(m => ({
     default: m.StateEventsDialogContent,
   })),
 )
