+++ src/components/app/pageEvents/common/stateEventsDialogContentSkeleton.tsx
