+++ src/components/app/pageEvents/common/successfulEventSignupDialog.tsx
