+++ src/components/app/pageEvents/common/upcomingEvents.tsx
@@ -3,7 +3,7 @@
 import { useMemo, useState } from 'react'
 import { isAfter } from 'date-fns'
 
-import { EventCard } from '@/components/app/pageEvents/components/eventCard'
+import { EventCard } from '@/components/app/pageEvents/common/eventCard'
 import { getUniqueEventKey } from '@/components/app/pageEvents/utils/getUniqueEventKey'
 import { Button } from '@/components/ui/button'
 import {
